+++ instrumentation/servlet/servlet-5.0/jetty11-testing/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v5_0/jetty/JettyServlet5SyncTest.java
@@ -7,19 +7,10 @@ package io.opentelemetry.javaagent.instrumentation.servlet.v5_0.jetty;
 
 import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.TestServlet5;
 import jakarta.servlet.Servlet;
-import org.junit.jupiter.api.condition.EnabledForJreRange;
-import org.junit.jupiter.api.condition.JRE;
 
-@EnabledForJreRange(min = JRE.JAVA_11)
-public class JettyServlet5AsyncTest extends JettyServlet5Test {
-
+public class JettyServlet5SyncTest extends JettyServlet5Test {
   @Override
   public Class<? extends Servlet> servlet() {
-    return TestServlet5.Async.class;
-  }
-
-  @Override
-  public boolean errorEndpointUsesSendError() {
-    return false;
+    return TestServlet5.Sync.class;
   }
 }
