+++ instrumentation/servlet/servlet-5.0/jetty11-testing/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v5_0/jetty/dispatch/JettyServlet5DispatchAsyncTest.java
@@ -18,43 +18,37 @@ import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint
 import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.TestServlet5;
 import jakarta.servlet.Servlet;
 import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.junit.jupiter.api.condition.EnabledForJreRange;
-import org.junit.jupiter.api.condition.JRE;
 
-@EnabledForJreRange(min = JRE.JAVA_11)
-public class JettyServlet5DispatchImmediateTest extends JettyDispatchTest {
+public class JettyServlet5DispatchAsyncTest extends JettyDispatchTest {
   @Override
   public Class<? extends Servlet> servlet() {
     return TestServlet5.Async.class;
   }
 
   @Override
-  public boolean errorEndpointUsesSendError() {
-    return false;
-  }
-
-  @Override
   protected void setupServlets(ServletContextHandler context) throws Exception {
     super.setupServlets(context);
     addServlet(
-        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), TestServlet5.DispatchImmediate.class);
+        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), TestServlet5.DispatchAsync.class);
     addServlet(
         context,
         "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
-        TestServlet5.DispatchImmediate.class);
-    addServlet(context, "/dispatch" + SUCCESS.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(context, "/dispatch" + QUERY_PARAM.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(context, "/dispatch" + ERROR.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(context, "/dispatch" + EXCEPTION.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(context, "/dispatch" + REDIRECT.getPath(), TestServlet5.DispatchImmediate.class);
+        TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + SUCCESS.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + QUERY_PARAM.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + ERROR.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + EXCEPTION.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + REDIRECT.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + AUTH_REQUIRED.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + CAPTURE_HEADERS.getPath(), TestServlet5.DispatchAsync.class);
     addServlet(
-        context, "/dispatch" + AUTH_REQUIRED.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(
-        context, "/dispatch" + CAPTURE_HEADERS.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(
-        context, "/dispatch" + CAPTURE_PARAMETERS.getPath(), TestServlet5.DispatchImmediate.class);
-    addServlet(
-        context, "/dispatch" + INDEXED_CHILD.getPath(), TestServlet5.DispatchImmediate.class);
+        context, "/dispatch" + CAPTURE_PARAMETERS.getPath(), TestServlet5.DispatchAsync.class);
+    addServlet(context, "/dispatch" + INDEXED_CHILD.getPath(), TestServlet5.DispatchAsync.class);
     addServlet(context, "/dispatch/recursive", TestServlet5.DispatchRecursive.class);
+  }
+
+  @Override
+  public boolean errorEndpointUsesSendError() {
+    return false;
   }
 }
