+++ instrumentation/servlet/servlet-5.0/jetty11-testing/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v5_0/jetty/dispatch/JettyServlet5ForwardTest.java
@@ -15,43 +15,42 @@ import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.REDIRECT;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.SUCCESS;
 
+import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.RequestDispatcherServlet;
 import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.TestServlet5;
 import jakarta.servlet.Servlet;
 import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.junit.jupiter.api.condition.EnabledForJreRange;
-import org.junit.jupiter.api.condition.JRE;
 
-@EnabledForJreRange(min = JRE.JAVA_11)
-public class JettyServlet5DispatchAsyncTest extends JettyDispatchTest {
+public class JettyServlet5ForwardTest extends JettyDispatchTest {
   @Override
   public Class<? extends Servlet> servlet() {
-    return TestServlet5.Async.class;
+    return TestServlet5.Sync.class; // dispatch to sync servlet
   }
 
   @Override
   protected void setupServlets(ServletContextHandler context) throws Exception {
     super.setupServlets(context);
+
+    addServlet(context, "/dispatch" + SUCCESS.getPath(), RequestDispatcherServlet.Forward.class);
     addServlet(
-        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), TestServlet5.DispatchAsync.class);
+        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), RequestDispatcherServlet.Forward.class);
     addServlet(
         context,
         "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
-        TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + SUCCESS.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + QUERY_PARAM.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + ERROR.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + EXCEPTION.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + REDIRECT.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + AUTH_REQUIRED.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + CAPTURE_HEADERS.getPath(), TestServlet5.DispatchAsync.class);
+        RequestDispatcherServlet.Forward.class);
     addServlet(
-        context, "/dispatch" + CAPTURE_PARAMETERS.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch" + INDEXED_CHILD.getPath(), TestServlet5.DispatchAsync.class);
-    addServlet(context, "/dispatch/recursive", TestServlet5.DispatchRecursive.class);
-  }
-
-  @Override
-  public boolean errorEndpointUsesSendError() {
-    return false;
+        context, "/dispatch" + QUERY_PARAM.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(context, "/dispatch" + REDIRECT.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(context, "/dispatch" + ERROR.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(context, "/dispatch" + EXCEPTION.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context, "/dispatch" + AUTH_REQUIRED.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context, "/dispatch" + CAPTURE_HEADERS.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context,
+        "/dispatch" + CAPTURE_PARAMETERS.getPath(),
+        RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context, "/dispatch" + INDEXED_CHILD.getPath(), RequestDispatcherServlet.Forward.class);
   }
 }
