+++ instrumentation/servlet/servlet-5.0/testing/build.gradle.kts
@@ -13,16 +13,12 @@ dependencies {
 
   testImplementation(project(":instrumentation:servlet:servlet-5.0:testing"))
 
-  // Jetty 11 requires Java 11
-  testLibrary("org.eclipse.jetty:jetty-server:11.0.0")
-  testLibrary("org.eclipse.jetty:jetty-servlet:11.0.0")
   testLibrary("org.apache.tomcat.embed:tomcat-embed-core:10.0.0")
   testLibrary("org.apache.tomcat.embed:tomcat-embed-jasper:10.0.0")
 
   // Tomcat 10.1 requires Java 11
   latestDepTestLibrary("org.apache.tomcat.embed:tomcat-embed-core:10.0.+")
   latestDepTestLibrary("org.apache.tomcat.embed:tomcat-embed-jasper:10.0.+")
-  latestDepTestLibrary("org.eclipse.jetty:jetty-server:11.+")
 }
 
 tasks {
