+++ instrumentation/servlet/servlet-5.0/testing/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v5_0/tomcat/dispatch/TomcatServlet5ForwardTest.java
@@ -18,8 +18,8 @@ import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerTestOptions;
+import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.RequestDispatcherServlet;
 import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.TestServlet5;
-import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.tomcat.RequestDispatcherServlet;
 import jakarta.servlet.Servlet;
 import org.apache.catalina.Context;
 import org.junit.jupiter.api.extension.RegisterExtension;
