+++ settings.gradle.kts
@@ -554,6 +554,7 @@ include(":instrumentation:servlet:servlet-3.0:javaagent-unit-tests")
 include(":instrumentation:servlet:servlet-3.0:testing")
 include(":instrumentation:servlet:servlet-5.0:javaagent")
 include(":instrumentation:servlet:servlet-5.0:javaagent-unit-tests")
+include(":instrumentation:servlet:servlet-5.0:jetty11-testing")
 include(":instrumentation:servlet:servlet-5.0:jetty12-testing")
 include(":instrumentation:servlet:servlet-5.0:testing")
 include(":instrumentation:servlet:servlet-common:bootstrap")
